+++ sentry-android-navigation/src/main/java/io/sentry/android/navigation/SentryNavigationListener.kt
@@ -108,7 +108,9 @@ constructor(
     arguments: Map<String, Any?>,
   ) {
     if (!isPerformanceEnabled) {
-      TracingUtils.startNewTrace(scopes)
+      if (!scopes.options.isEnableSessionTraceLifecycle) {
+        TracingUtils.startNewTrace(scopes)
+      }
       return
     }