+++ sentry/src/test/java/io/sentry/SentryOptionsTest.kt
@@ -50,6 +50,11 @@ class SentryOptionsTest {
   }
 
   @Test
+  fun `when options is initialized, session trace lifecycle is disabled`() {
+    assertFalse(SentryOptions().isEnableSessionTraceLifecycle)
+  }
+
+  @Test
   fun `when options is initialized, integrations contain UncaughtExceptionHandlerIntegration`() {
     assertTrue(SentryOptions().integrations.any { it is UncaughtExceptionHandlerIntegration })
   }
@@ -420,6 +425,7 @@ class SentryOptionsTest {
     externalOptions.profileSessionSampleRate = 0.8
     externalOptions.profilingTracesDirPath = "/profiling-traces"
     externalOptions.profileLifecycle = ProfileLifecycle.TRACE
+    externalOptions.isEnableSessionTraceLifecycle = true
 
     val hash = StringUtils.calculateStringHash(externalOptions.dsn, mock())
     val options = SentryOptions()
@@ -484,6 +490,7 @@ class SentryOptionsTest {
     assertEquals(0.8, options.profileSessionSampleRate)
     assertEquals("/profiling-traces${File.separator}${hash}", options.profilingTracesDirPath)
     assertEquals(ProfileLifecycle.TRACE, options.profileLifecycle)
+    assertTrue(options.isEnableSessionTraceLifecycle)
   }
 
   @Test